Use int or int32_t instead of size_t where suitable

+++ src/BtPieceMessageValidator.cc
@@ -39,7 +39,7 @@
 namespace aria2 {
 
 BtPieceMessageValidator::BtPieceMessageValidator
-(const BtPieceMessage* message, size_t numPiece, size_t pieceLength)
+(const BtPieceMessage* message, size_t numPiece, int32_t pieceLength)
   : message_(message),
     numPiece_(numPiece),
     pieceLength_(pieceLength)
